requirements.txt
internetarchive>=5.5.1
PyYAML>=6.0
pytest>=7.0.0  # For running tests
